Improvement to patched function call
