+++ tests/py/mdb_test.py
@@ -13,63 +13,10 @@ from soprano.properties.nmr import MSIsotropy
 
 import mongomock
 # from mongomock.gridfs import enable_gridfs_integration
-import gridfs
-from unittest import mock
 
 file_path = os.path.split(__file__)[0]
 data_path = os.path.join(file_path, '../data')
 sys.path.append(os.path.abspath(os.path.join(file_path, '../../')))
-
-class MockGridFile:
-    def __init__(self, data):
-        # Store data as bytes. If data is a string, encode it to bytes using UTF-8.
-        self.data = data.encode('utf-8') if isinstance(data, str) else data
-
-    def read(self):
-        # Return the stored data.
-        return self.data
-
-class MockGridFS:
-    def __init__(self, database, collection='fs'):
-        # Initialize with a reference to the database and collection name.
-        self.database = database
-        self.collection = collection
-        # Dictionary to store files with their ObjectId as keys.
-        self.files = {}
-
-    def put(self, data, **kwargs):
-        # Generate a new ObjectId for the file.
-        file_id = ObjectId()
-        # Store the data in the files dictionary.
-        self.files[file_id] = data
-        # Return the ObjectId of the stored file.
-        return file_id
-
-    def get(self, file_id):
-        # Retrieve the data associated with the given file_id.
-        data = self.files.get(file_id)
-        # Return a MockGridFile instance containing the data.
-        return MockGridFile(data)
-
-class MockGridIn:
-    def __init__(self, database, collection='fs'):
-        # Initialize with a reference to the database and collection name.
-        self.database = database
-        self.collection = collection
-
-class MockGridOut:
-    def __init__(self, database, collection='fs'):
-        # Initialize with a reference to the database and collection name.
-        self.database = database
-        self.collection = collection
-
-def enable_gridfs_integration():
-    # Patch the gridfs.GridFS class to use the MockGridFS class.
-    mock.patch('gridfs.GridFS', MockGridFS).start()
-    # Patch the gridfs.GridIn class to use the MockGridIn class.
-    mock.patch('gridfs.GridIn', MockGridIn).start()
-    # Patch the gridfs.GridOut class to use the MockGridOut class.
-    mock.patch('gridfs.GridOut', MockGridOut).start()
 
 _fake_orcid = {
     'path': '0000-0000-0000-0000',
@@ -103,6 +50,7 @@ class MagresDBTest(unittest.TestCase):
     @mongomock.patch("mongodb://localhost:27017", on_new="pymongo")
     def setUp(self):
 
+        from ccpncdb.gridfs_patch import enable_gridfs_integration
         enable_gridfs_integration()
         
         from ccpncdb.config import Config
