+++ tests/py/server_test.py
@@ -14,7 +14,7 @@ from soprano.selection import AtomSelection
 from soprano.properties.nmr import MSIsotropy
 
 import mongomock
-from mongomock.gridfs import enable_gridfs_integration
+# from mongomock.gridfs import enable_gridfs_integration
 
 file_path = os.path.split(__file__)[0]
 data_path = os.path.join(file_path, '../data')
@@ -46,6 +46,8 @@ class MDBServerTest(unittest.TestCase):
 
     @mongomock.patch("mongodb://localhost:27017", on_new="pymongo")
     def setUp(self):
+        
+        from ccpncdb.gridfs_patch import enable_gridfs_integration
         enable_gridfs_integration()
 
         from ccpncdb.server import MainServer
